use mysqli multi_query for importing dump

+++ src/QtDatabase.php
@@ -163,8 +163,11 @@ class QtDatabase
 
             $this->writeLn('Downloading dump file');
 
-            if (!($ftpServer->downloadFile($dataFile, $dumpFile.'.aes') && $extension = 'aes') && !($ftpServer->downloadFile($dataFile, $dumpFile.'.box') && $extension = 'box')) {
-                throw new \Exception('Could not download '.$dumpFile);
+            if (
+                !($ftpServer->downloadFile($dataFile, $dumpFile.'.box') && $extension = 'box') &&
+                !($ftpServer->downloadFile($dataFile, $dumpFile.'.aes') && $extension = 'aes')
+            ) {
+                throw new \Exception('Could not download '.$dumpFile.'.'.$extension);
             }
 
             $this->writeLn('Downloading key file');
@@ -197,9 +200,13 @@ class QtDatabase
             $this->writeLn('Extracting zip file');
 
             $zip = new \ZipArchive();
-            if (!$zip->open($zipFile) || !($stat = @$zip->statIndex(0)) || !($stream = @$zip->getStream($zip->getNameIndex(0)))) {
+
+            if (true !== $zip->open($zipFile)) {
                 throw new \Exception('Error reading the zip file.');
             }
+
+            $zip->extractTo('/tmp');
+            $zip->close();
 
             $this->writeLn('Dropping and re-creating database');
             if ($this->doesExist() && !$this->getAdminConnection()->query('DROP DATABASE '.$this->config['database'])) {
@@ -213,7 +220,7 @@ class QtDatabase
             } catch (\Exception $exception) {
                 // Fail silently
             }
-            $conn = @mysqli_connect($this->config['hostspec'], $this->config['username'], $this->config['password']);
+            $conn = $this->getAdminConnection();
             $conn->select_db($this->config['database']);
             @$conn->query('set sql_mode=NO_ENGINE_SUBSTITUTION,innodb_strict_mode=0');
 
@@ -221,29 +228,10 @@ class QtDatabase
             $delimiter = ';';
 
             $this->writeLn('Importing dump');
-            while (!feof($stream)) {
-                $line = trim(fgets($stream), "\t\n\r\0");
-                if ('--' != substr($line, 0, 2)) {
-                    if ('' != $statement && ' ' != substr($statement, -1) && $line && ' ' != $line[0]) {
-                        $statement .= ' ';
-                    }
-                    $statement .= $line;
-                    if (preg_match("'^DELIMITER (.+)$'", $statement, $matches)) {
-                        $delimiter = $matches[1];
-                        $statement = '';
-                    } elseif (preg_match('/'.str_replace(';', '\;', $delimiter).'$/', $statement)) {
-                        // Filter definer information
-                        $statement = preg_replace('/ DEFINER=`[^`]+`@`[^`]+`/', '', $statement);
-                        // Filter database name
-                        $statement = preg_replace('/ALTER DATABASE `[a-z0-9_]+`/i', 'ALTER DATABASE', $statement);
-                        if (!$conn->query($statement)) {
-                            throw new \Exception($conn->error.' ('.$statement.')');
-                        }
-                        $statement = '';
-                    }
-                }
+
+            if (!$conn->multi_query(file_get_contents($sqlFile = "/tmp/{$dumpFile}.sql"))) {
+                throw new \Exception('Database import failed');
             }
-            fclose($stream);
 
             $conn->query('DROP TABLE IF EXISTS dbrevision;');
             $conn->query('CREATE TABLE dbrevision (branch TEXT, revision varchar(255), masked tinyint not null default 0);');
@@ -270,11 +258,13 @@ class QtDatabase
             @unlink($dataFile);
             @unlink($keyFile);
             @unlink($zipFile);
+            @unlink($sqlFile);
             throw $e;
         }
         @unlink($dataFile);
         @unlink($keyFile);
         @unlink($zipFile);
+        @unlink($sqlFile);
     }
 
     public function killProcess($id)
